Lab 09-- created movie.js, weather.js files and moved functionality from server.js. Server is live and running----- :)
